Some fixes after merge
