+++ src/main/java/com/example/thymeleaf/service/TeamService.java
@@ -1,0 +1,2 @@
+package com.example.thymeleaf.service;public interface TeamService {
+}
