+++ src/main/java/com/example/thymeleaf/service/TeamServiceImpl.java
@@ -1,0 +1,2 @@
+package com.example.thymeleaf.service;public class TeamServiceImpl {
+}
